Added a new Integration Test project, added a couple SignalR tests as well, and fixed a couple errors

+++ VoterSystem.Shared/Dto/ChoiceResultDto.cs
@@ -5,5 +5,5 @@ namespace VoterSystem.Shared.Dto;
 public class ChoiceResultDto
 {
     public required long ChoiceId { get; init; }
-    [Range(0, long.MaxValue)] public required long VoteCount { get; init; }
+    [Range(0, int.MaxValue)] public required long VoteCount { get; init; }
 }
